organized/document/requirements.txt
scikit-learn>=1.0.2
Pillow>=9.0.0
transformers>=4.30.0
PyPDF2>=3.0.0
click>=8.1.3
rich>=13.0.0
numpy>=1.24.0
torch>=2.0.0
python-magic-bin>=0.4.14
tqdm>=4.65.0 